feat:create repositories.

+++ src/main/java/com/mvc/repository/OrderProductRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderProductRepository extends JpaRepository<OrderProduct, Integer> {
+public interface OrderProductRepository extends JpaRepository<OrderProduct, Long> {
 }
